Put the create command into its final resting place.
